+++ weblate/lang/forms.py
@@ -16,8 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
-
-
+from crispy_forms.helper import FormHelper
 from django import forms
 
 from weblate.lang.models import Language, Plural
@@ -28,6 +27,11 @@ class LanguageForm(forms.ModelForm):
         model = Language
         exclude = []
 
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.helper = FormHelper()
+        self.helper.form_tag = False
+
     @staticmethod
     def get_field_doc(field):
         return ("admin/languages", "language-{}".format(field.name))
@@ -37,6 +41,11 @@ class PluralForm(forms.ModelForm):
     class Meta:
         model = Plural
         fields = ["number", "formula"]
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.helper = FormHelper()
+        self.helper.form_tag = False
 
     @staticmethod
     def get_field_doc(field):