+++ weblate/trans/guide.py
@@ -55,7 +55,7 @@ class Guideline:
             url = "{}#{}".format(url, self.anchor)
         return url
 
-    def get_docs_url(self):
+    def get_doc_url(self, user=None):
         return ""
 
 
@@ -75,8 +75,8 @@ class Group(Guideline):
 class VCSGroup(Group):
     description = _("Version control integration")
 
-    def get_docs_url(self):
-        return get_doc_url("vcs")
+    def get_doc_url(self, user=None):
+        return get_doc_url("vcs", user=user)
 
 
 @register
@@ -91,10 +91,10 @@ class HookGuideline(Guideline):
         return self.component.change_set.filter(action=Change.ACTION_HOOK).exists()
 
     def get_url(self):
-        return self.get_docs_url()
+        return self.get_doc_url()
 
-    def get_docs_url(self):
-        return get_doc_url("admin/continuous", "update-vcs")
+    def get_doc_url(self, user=None):
+        return get_doc_url("admin/continuous", "update-vcs", user=user)
 
 
 @register
@@ -108,16 +108,16 @@ class PushGuideline(Guideline):
     def is_passing(self):
         return self.component.can_push()
 
-    def get_docs_url(self):
-        return get_doc_url("admin/continuous", "push-changes")
+    def get_doc_url(self, user=None):
+        return get_doc_url("admin/continuous", "push-changes", user=user)
 
 
 @register
 class CommunityGroup(Group):
     description = _("Building community")
 
-    def get_docs_url(self):
-        return get_doc_url("devel/community")
+    def get_doc_url(self, user=None):
+        return get_doc_url("devel/community", user=user)
 
 
 @register
@@ -132,8 +132,8 @@ class InstructionsGuideline(Guideline):
             "settings", kwargs=self.component.project.get_reverse_url_kwargs()
         )
 
-    def get_docs_url(self):
-        return get_doc_url("admin/project", "project")
+    def get_doc_url(self, user=None):
+        return get_doc_url("admin/project", "project", user=user)
 
 
 @register
@@ -145,7 +145,7 @@ class LicenseGuideline(Guideline):
     def is_passing(self):
         return self.component.libre_license
 
-    def get_docs_url(self):
+    def get_doc_url(self, user=None):
         return "https://choosealicense.com/"
 
 
@@ -158,16 +158,16 @@ class AlertGuideline(Guideline):
     def is_passing(self):
         return not self.component.all_alerts
 
-    def get_docs_url(self):
-        return get_doc_url("devel/alerts")
+    def get_doc_url(self, user=None):
+        return get_doc_url("devel/alerts", user=user)
 
 
 @register
 class ContextGroup(Group):
     description = _("Provide context to the translators")
 
-    def get_docs_url(self):
-        return get_doc_url("admin/translating", "additional")
+    def get_doc_url(self, user=None):
+        return get_doc_url("admin/translating", "additional", user=user)
 
 
 @register
@@ -180,8 +180,8 @@ class ScreenshotGuideline(Guideline):
 
         return Screenshot.objects.filter(translation__component=self.component).exists()
 
-    def get_docs_url(self):
-        return get_doc_url("admin/translating", "screenshots")
+    def get_doc_url(self, user=None):
+        return get_doc_url("admin/translating", "screenshots", user=user)
 
 
 @register
@@ -198,8 +198,8 @@ class FlagsGuideline(Guideline):
             ).exists()
         )
 
-    def get_docs_url(self):
-        return get_doc_url("admin/checks", "custom-checks")
+    def get_doc_url(self, user=None):
+        return get_doc_url("admin/checks", "custom-checks", user=user)
 
 
 @register
@@ -221,16 +221,16 @@ class SafeHTMLGuideline(Guideline):
             ).exists()
         )
 
-    def get_docs_url(self):
-        return get_doc_url("user/checks", "check-safe-html")
+    def get_doc_url(self, user=None):
+        return get_doc_url("user/checks", "check-safe-html", user=user)
 
 
 @register
 class AddonsGroup(Group):
     description = _("Workflow customization")
 
-    def get_docs_url(self):
-        return get_doc_url("admin/addons")
+    def get_doc_url(self, user=None):
+        return get_doc_url("admin/addons", user=user)
 
 
 class AddonGuideline(Guideline):
@@ -250,8 +250,10 @@ class AddonGuideline(Guideline):
         addon = ADDONS[self.addon]
         return addon.can_install(self.component, None)
 
-    def get_docs_url(self):
-        return get_doc_url("admin/addons", ADDONS[self.addon].get_doc_anchor())
+    def get_doc_url(self, user=None):
+        return get_doc_url(
+            "admin/addons", ADDONS[self.addon].get_doc_anchor(), user=user
+        )
 
     @property
     def description(self):