+++ docs/changes.rst
@@ -5,6 +5,7 @@ Not yet released.
 
 * Improved automatic translation performance.
 * Fixed session expiry for authenticated users.
+* Add support for hiding version information.
 
 Weblate 4.3
 -----------